Final Implementation of Sign up and Log in with simple designing
