+++ templates/dashboard.html
@@ -19,7 +19,7 @@
             <li><a href="#elections">Elections</a></li>
             <li><a href="#vote-count">Vote Count</a></li>
             <li><a href="#candidate">Candidate</a></li>
-            <li><a href="#login">Login</a></li>
+           
         </ul>
     </nav>
 
@@ -56,11 +56,7 @@
             <h2>Candidate</h2>
             <p>Explore candidate profiles and their manifestos.</p>
         </div>
-        <div class="service-card" id="login">
-            <i class="fas fa-sign-in-alt"></i>
-            <h2>Login</h2>
-            <p>Access your account to cast your vote securely.</p>
-        </div>
+        
     </section>
 
 </body>
