+++ templates/signup.html
@@ -4,20 +4,18 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Login</title>
-    <link rel="stylesheet" href="{% static 'css/styles.css' %}">
+    <title>Sign Up</title>
+    <link rel="stylesheet" href="{% static 'css/style.css' %}">
 </head>
 <body class="sign-in-page">
     <div class="container">
-        <h2>Login</h2>
+        <h2>Sign Up</h2>
         <form method="POST">
             {% csrf_token %}
-            {{ form.as_p }}
-            <button type="submit">Login</button>
+            {{ form.as_p }} 
+            <button type="submit">Register</button>
         </form>
-        <p>Don't have an account? <a href="{% url 'signup' %}">Sign Up</a></p>
+        <p>Already have an account? <a href="{% url 'login' %}">Login</a></p>
     </div>
 </body>
 </html>
-
-
