+++ voting_site/settings.py
@@ -32,15 +32,14 @@ ALLOWED_HOSTS = []
 
 INSTALLED_APPS = [
      
-    'users',
-    'ovs',
+    
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-   
+    'ovs',
 ]
 
 MIDDLEWARE = [
@@ -88,7 +87,7 @@ DATABASES = {
 # Password validation
 # https://docs.djangoproject.com/en/5.1/ref/settings/#auth-password-validators
 
-AUTH_USER_MODEL = 'users.User_db'
+
 
 AUTH_PASSWORD_VALIDATORS = [
     {
